Move OrcToIcebergVisitor to separate file
